requirements.txt
# requirements.txt

annotated-types==0.7.0
anyio==4.7.0
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.1
python-multipart==0.0.19
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1

# MySQL
pymysql==1.1.0
cryptography==42.0.8

# Authentication
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
email-validator==2.1.1
